[configure] Fix automake version check.
The configure script fails on automake-1.11 (in Fedora-11) since
it looks for 3-digit version number.

Patch fixed by "Yann E. MORIN", with the added comment:

The check for the automake version is not against a 3-digit number,
but really against a 3-part version number, a-la 'x.y.z'. Versions
such as 1.10 and 1.11 are also valid.