libc/glibc: workaround for autoconf'ed define of caddr_t
When building canadian cross compiler, I have some trouble with
configure defining caddr_t as a macro, like:
  #define caddr_t char *

When combined with the types.h where caddr_t is protected together
with daddr_t, the typedef of caddr_t breaks.

This patch works around it by protecting the caddr_t typedef
specifically.

I am uncertain as to the real cause and solution to this :-(

Signed-off-by: Esben Haabendal <eha@dev.doredevelopment.dk>