docs: --target versus --host
In the very beginnings, eons ago, autotools also got confused by this
whole build vs. host vs. target, and got it wrong. Now they fixed it,
but they want to keep backward compatibility, so the --target is still
recongised, although ./configure will complain if you do so. It is
better to use --host.

Signed-off-by: "Trevor Woerner" <twoerner@gmail.com>
[yann.morin.1998@anciens.enib.fr: add build/host clarification]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>